# 8357959: (bf) ByteBuffer.allocateDirect initialization can result in large TTSP spikes