--- a/test/jdk/java/nio/Buffer/AllocateDirectInit.java
+++ b/test/jdk/java/nio/Buffer/AllocateDirectInit.java
@@ -23,29 +23,60 @@
 
 /**
  * @test
- * @bug 4490253 6535542
+ * @bug 4490253 6535542 8357959
+ * @key randomness
+ * @library /test/lib
+ * @build jdk.test.lib.RandomFactory
  * @summary Verify that newly allocated direct buffers are initialized.
+ * @run main/othervm AllocateDirectInit
  */
 
 import java.nio.ByteBuffer;
+import java.util.Random;
+
+import jdk.test.lib.RandomFactory;
 
 public class AllocateDirectInit {
+    private static final int MAX_BIN_LIMIT = 16 * 1024 * 1024;
+    private static final int MAX_DEC_LIMIT = 10 * 1000 * 1000;
+    private static final int TRIES_PER_LIMIT = 1024;
+
+    private static final Random RND = RandomFactory.getRandom();
+
     public static void main(String [] args){
-        for (int i = 0; i < 1024; i++) {
-            ByteBuffer bb = ByteBuffer.allocateDirect(1024);
-//          printByteBuffer(bb);
-            for (bb.position(0); bb.position() < bb.limit(); ) {
-                if ((bb.get() & 0xff) != 0)
-                    throw new RuntimeException("uninitialized buffer, position = "
-                                               + bb.position());
+        // Try power of two limits
+        for (int limit = 1; limit < MAX_BIN_LIMIT; limit *= 2) {
+            check(ByteBuffer.allocateDirect(limit - 1));
+            check(ByteBuffer.allocateDirect(limit));
+            check(ByteBuffer.allocateDirect(limit + 1));
+        }
+
+        // Try power of ten limits
+        for (int limit = 1; limit < MAX_DEC_LIMIT; limit *= 10) {
+            check(ByteBuffer.allocateDirect(limit - 1));
+            check(ByteBuffer.allocateDirect(limit));
+            check(ByteBuffer.allocateDirect(limit + 1));
+        }
+
+        // Try random sizes within power of two limits
+        for (int limit = 1; limit < MAX_BIN_LIMIT; limit *= 2) {
+            for (int t = 0; t < TRIES_PER_LIMIT; t++) {
+                check(ByteBuffer.allocateDirect(RND.nextInt(limit)));
             }
         }
     }
 
-    private static void printByteBuffer(ByteBuffer bb) {
-        System.out.print("byte [");
-        for (bb.position(0); bb.position() < bb.limit(); )
-            System.out.print(" " + Integer.toHexString(bb.get() & 0xff));
-        System.out.println(" ]");
+    private static void check(ByteBuffer bb) {
+        for (bb.position(0); bb.position() < bb.limit(); ) {
+            if (bb.get() != 0) {
+                int mismatchPos = bb.position();
+                System.out.print("byte [");
+                for (bb.position(0); bb.position() < bb.limit(); ) {
+                    System.out.print(" " + Integer.toHexString(bb.get() & 0xff));
+                }
+                System.out.println(" ]");
+                throw new RuntimeException("uninitialized buffer, position = " + mismatchPos);
+            }
+        }
     }
 }