--- a/test/jdk/java/nio/Buffer/AllocateDirectInit.java
+++ b/test/jdk/java/nio/Buffer/AllocateDirectInit.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2001, 2007, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2001, 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -67,7 +67,7 @@ public class AllocateDirectInit {
     }
 
     private static void check(ByteBuffer bb) {
-        for (bb.position(0); bb.position() < bb.limit(); ) {
+        while (bb.hasRemaining()) {
             if (bb.get() != 0) {
                 int mismatchPos = bb.position();
                 System.out.print("byte [");