--- a/test/micro/org/openjdk/bench/java/nio/ByteBufferAllocationBenchmark.java
+++ b/test/micro/org/openjdk/bench/java/nio/ByteBufferAllocationBenchmark.java
@@ -39,8 +39,9 @@ import org.openjdk.jmh.annotations.Warmup;
 @BenchmarkMode(Mode.AverageTime)
 @OutputTimeUnit(TimeUnit.NANOSECONDS)
 @State(Scope.Thread)
-@Measurement(iterations = 5)
-@Warmup(iterations = 5)
+@Warmup(iterations = 3, time = 1, timeUnit = TimeUnit.SECONDS)
+@Measurement(iterations = 3, time = 1, timeUnit = TimeUnit.SECONDS)
+@Fork(3)
 public class ByteBufferAllocationBenchmark {
 
     @Param({