--- a/test/micro/org/openjdk/bench/java/nio/DirectByteBufferAlloc.java
+++ b/test/micro/org/openjdk/bench/java/nio/DirectByteBufferAlloc.java
@@ -28,6 +28,7 @@ import java.nio.ByteBuffer;
 import java.util.concurrent.TimeUnit;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
+import org.openjdk.jmh.annotations.Fork;
 import org.openjdk.jmh.annotations.Measurement;
 import org.openjdk.jmh.annotations.Mode;
 import org.openjdk.jmh.annotations.OutputTimeUnit;
@@ -42,13 +43,13 @@ import org.openjdk.jmh.annotations.Warmup;
 @Warmup(iterations = 3, time = 1, timeUnit = TimeUnit.SECONDS)
 @Measurement(iterations = 3, time = 1, timeUnit = TimeUnit.SECONDS)
 @Fork(3)
-public class ByteBufferAllocationBenchmark {
+public class DirectByteBufferAlloc {
 
     @Param({
             "128", // 128 bytes
             "1024", // 1KB
             "1048576", // 1 MB
-            "2147483647" // ~2 GB
+            "16777216" // 16MB
     })
     public int bytes;