ByteBuffer.allocateDirect uses UNSAFE.setMemory, causing high time-to-safepoint (100+ ms) for large (100 MB+) allocations.

This PR applies a simple fix by chunking the zeroing operation within ByteBuffers. A more robust solution would be to add chunking inside UNSAFE.setMemory itself. However Its not that straightforward as mentioned by Aleksey in [JDK-8357959](https://bugs.openjdk.org/browse/JDK-8357959)
>Looks like all current uses we care about are in Buffers. Taking a safepoint within cleaning would open some questions whether any VM code expect to see semi-initialized area we are busy cleaning up. For Buffers, this question does not arise. Therefore, we can do the fix in Buffers first, without changing the Unsafe itself.
 
I can pursue that if its preferred. I chose 1 MB as a chunk size some what arbitrarily I am open to suggestion, if there are better options.

For verification, I tested the fix against the reproducer - [gist](https://gist.github.com/rk-kmr/be4322b72a14ae04aeefc0260c01acf6) and confirmed that ttsp timing were lower.

**before**
```
0.444s][info][safepoint,stats] ThreadDump                   [             13               1 ][        194156625      65291  194221916 ]               0
[0.662s][info][safepoint,stats] ThreadDump                   [             13               1 ][        200013875      87834  200101709 ]               0
[0.858s][info][safepoint,stats] ThreadDump                   [             13               1 ][        183762583      43417  183806000 ]               0
[1
```
**after**
```
1.705s][info][safepoint,stats] ThreadDump                   [             11               1 ][            92792      24958     117750 ]               0
[1.724s][info][safepoint,stats] ThreadDump                   [             11               1 ][           497375      94041     591416 ]               0
[1.736s][info][safepoint,stats] ThreadDump                   [             11               1 ][           156750      47208     203958 ]               0
[1.747s][info][safepoint,stats] ThreadDump                   [             11               1 ][           121958      28334     150292 ]               0

```
I added a benchmark to ensure that chunking doesn't introduce significant overhead across different allocation sizes, and following results confirm that. 

**Before**
```
Benchmark                                              (bytes)  Mode  Cnt          Score         Error  Units
ByteBufferAllocationBenchmark.allocateDirectBuffer         128  avgt   25        860.869 ±     170.921  ns/op
ByteBufferAllocationBenchmark.allocateDirectBuffer        1024  avgt   25       1368.599 ±     320.159  ns/op
ByteBufferAllocationBenchmark.allocateDirectBuffer     1048576  avgt   25      80169.230 ±     448.075  ns/op
ByteBufferAllocationBenchmark.allocateDirectBuffer  2147483647  avgt   25  635047372.458 ± 5461328.161  ns/op
```
**After**
```
Benchmark                                              (bytes)  Mode  Cnt          Score         Error  Units
ByteBufferAllocationBenchmark.allocateDirectBuffer         128  avgt   25        869.676 ±     160.633  ns/op
ByteBufferAllocationBenchmark.allocateDirectBuffer        1024  avgt   25       1328.623 ±     282.198  ns/op
ByteBufferAllocationBenchmark.allocateDirectBuffer     1048576  avgt   25      80224.157 ±     454.906  ns/op
ByteBufferAllocationBenchmark.allocateDirectBuffer  2147483647  avgt   25  667037308.198 ± 4818648.535  ns/op
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] Change must be properly reviewed (2 reviews required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer), 1 [Author](https://openjdk.org/bylaws#author))

### Issue
 * [JDK-8357959](https://bugs.openjdk.org/browse/JDK-8357959): (bf) ByteBuffer.allocateDirect initialization can result in large TTSP spikes (**Enhancement** - P4)


### Reviewers
 * [Aleksey Shipilev](https://openjdk.org/census#shade) (@shipilev - **Reviewer**)
 * [Alan Bateman](https://openjdk.org/census#alanb) (@AlanBateman - **Reviewer**)

### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/25487/head:pull/25487` \
`$ git checkout pull/25487`

Update a local copy of the PR: \
`$ git checkout pull/25487` \
`$ git pull https://git.openjdk.org/jdk.git pull/25487/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 25487`

View PR using the GUI difftool: \
`$ git pr show -t 25487`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/25487.diff">https://git.openjdk.org/jdk/pull/25487.diff</a>

</details>
<details><summary>Using Webrev</summary>

[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/25487#issuecomment-2916093452)
</details>