@AlanBateman 2025-05-30
Moving it to Bits.setMemory looks good.