@shipilev 2025-06-04
I think this looks fine. @AlanBateman -- do you think it is worth having in JDK 25? If so, we need to integrate this soon.