@tstuefe 2025-06-04
This may be a stupid question but why do we not just mmap large allocations in Unsafe.allocateMemory (with anonymous mappings)? These sections would then come pre-zero-initialized by the OS, no?