@mcimadamore 2025-06-04
Note: in FFM we actually use a plain `long` loop instead of calling `Unsafe::setMemory`:
https://github.com/openjdk/jdk/blob/master/src/java.base/share/classes/jdk/internal/foreign/SegmentFactories.java#L239

(to facilitate this, we ensure that all allocations are, internally, at least 8-byte aligned).

A simple loop like this, with only one input region of memory typically auto-vectorizes quite well. In the long run, I'd prefer to see bulk operations in Unsafe being rewritten to use simpler operations and loops. I think we're getting quite close, but we can't get all the way there for stuff like `copyMemory` because, currently, there are some problems in determining that two off-heap memory regions are provably disjoint (but this is not an issue in the case of `setMemory`).