@mcimadamore 2025-06-04
For the records, while the fix looks fine, I'm a bit on the fence whether it's too late to target 25, as sometimes these things can hit unexpected performance regressions with more real-world/complex code.