@AlanBateman 2025-06-04
> I think this looks fine. @AlanBateman -- do you think it is worth having in JDK 25? I do. If you agree, we need to integrate this soon.

I don't see anything wrong with the current change but it does feel like a change for early in a release rather than the day before forking.  Also this issue has existed for 20+ years so maybe better to hold off to after the fork.