@AlanBateman 2025-06-10
> Right, makes sense. I think we are now early in a JDK 26 release.

Indeed, I'll do another pass on this today.